# docs/openapi.json
{
  "openapi": "3.0.3",
  "paths": {
...
    "/tickets": {
      "post": {
        "operationId": "TicketsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ticket created."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a ticket for the authenticated requester",
        "tags": [
          "tickets"
        ]
      },
      "get": {
        "operationId": "TicketsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "IN_PROGRESS",
                "WAITING_USER",
                "RESOLVED",
                "CLOSED"
              ]
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "LOW",
                "MEDIUM",
                "HIGH",
                "CRITICAL"
              ]
            }
          },
          {
            "name": "assigneeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated tickets."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List visible tickets with pagination, search and filters",
        "tags": [
          "tickets"
        ]
      }
    },
    "/tickets/{id}": {
      "get": {
        "operationId": "TicketsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a visible ticket",
        "tags": [
          "tickets"
        ]
      },
      "patch": {
        "operationId": "TicketsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update workflow, priority or assignment",
        "tags": [
          "tickets"
        ]
      }
    },
    "/tickets/{id}/comments": {
      "post": {
        "operationId": "TicketsController_addComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a public or authorized internal comment",
        "tags": [
          "tickets"
        ]
      },
      "get": {
        "operationId": "TicketsController_listComments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 10000,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 50,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List visible comments for a ticket",
        "tags": [
          "tickets"
        ]
      }
    },
    "/tickets/{id}/watchers": {
      "post": {
        "operationId": "TicketsController_addWatcher",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTicketWatcherDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add an active user as a ticket watcher",
        "tags": [
          "tickets"
        ]
      }
    },
    "/tickets/{id}/watchers/{userId}": {
      "delete": {
        "operationId": "TicketsController_removeWatcher",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a ticket watcher",
        "tags": [
          "tickets"
        ]
      }
    },
...
  },
  "info": {
    "title": "Secure Service Desk API",
    "description": "Secure, modular service desk API for portfolio demonstration. The MVP is single-tenant and manages identity, tickets, files, audit, notifications and reports in code.",
    "version": "0.1.0",
    "contact": {},
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/license/mit/"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Local demo reverse proxy"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Standalone API process"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "service_desk_refresh"
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-csrf-token",
        "description": "Double-submit CSRF token returned by register or login."
      }
    },
    "schemas": {
...
      "CreateTicketDto": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string",
            "example": "Cannot access the customer portal",
            "maxLength": 160
          },
          "description": {
            "type": "string",
            "example": "The portal returns an unauthorized response after signing in.",
            "maxLength": 10000
          }
        },
        "required": [
          "subject",
          "description"
        ]
      },
      "Object": {
        "type": "object",
        "properties": {}
      },
      "UpdateTicketDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "description": "Current ticket version used for optimistic concurrency control.",
            "minimum": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "WAITING_USER",
              "RESOLVED",
              "CLOSED"
            ]
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ]
          },
          "assigneeId": {
            "type": "object",
            "description": "MongoDB user ID. Send null to remove the assignee.",
            "nullable": true
          },
          "resolution": {
            "type": "string",
            "description": "Required when resolving a ticket.",
            "minLength": 5,
            "maxLength": 5000
          }
        },
        "required": [
          "version"
        ]
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "maxLength": 10000
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "INTERNAL"
            ],
            "default": "PUBLIC",
            "description": "INTERNAL comments require SUPPORT or ADMIN."
          }
        },
        "required": [
          "body"
        ]
      },
      "AddTicketWatcherDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "Active user to add as a ticket watcher."
          }
        },
        "required": [
          "userId"
        ]
      },
...
    }
  }
}